worked on crafting system
created crafting menu widget. created workshed; workshed can be entered and has a work bench which when clicked on opens the crafting menu widget and escape exits, within the crafting menu widget we have a recipe list and an expand area for the recipe image, crafting costs, and description, the inventory slot struct now has a map for a crafting recipe that is an item id and a count. some misc. fixes