continued work on crafting system
created cancel button functionality: closes menu. created recipe button functionality: puts the image in the expanded image on the right, generates the resource list, fills in the description. the resource list will be red if you don't have enough of that item. the resource list is a picture of the item, the name, and the count required. the recipe will be red if any resource is lacking. started work on the craft button but have not finished