crafting system should be finished
The crafting button attemps to craft your selected item if there is no room in the inventory it does nothing. if the room is cleared by the resources being removed it is crafted. tested and can craft multiple. refreshes recipes and resource list to check if still can craft after craft. seems to work

bugs:
if you click out side of the chest or inventory buttons the widget loses focus and escape no longer functions to exit

To do:
implement MF levels and gate
functionality
more recipes
able to place down chest