Save Implemented
I am unsure of what was specifically done in the last few commits but overall. A save system that keeps track of what was where in each level between travelling. If the player sleeps it is written to their drive, able to be loaded from that point upon exiting. Otherwise all progress for the day is lost. Implemented a day/night cycle. You pass out at 2 AM and when your stamina or health is exhausted.